@@ -1,58 +1,59 @@
 {
-	"name": "@micro-fleet/id-generator",
-	"description": "Provides methods to generate database primary key ID.",
-	"version": "1.0.2",
-	"scripts": {
-		"build": "npm run compile && npm run test",
-		"clean": "rimraf dist coverage .nyc_output",
-		"definition": "node ../../../scripts/definition.js",
-		"dedup": "rimraf node_modules/**/node_modules/**/typings",
-		"compile": "npm run clean && npm run tslint && tsc --pretty",
-		"release": "npm run build && npm run definition",
-		"test": "nyc --reporter=json --reporter=html --include dist/app/**/*.js mocha --recursive dist/test",
-		"tslint": "tslint -c tslint.json 'src/**/*.ts'",
-		"prepublishOnly": "npm run release",
-		"watch": "npm run clean && npm run tslint && tsc --watch --pretty"
-	},
-	"dependencies": {
-		"flake-idgen": "~1.1",
-		"int64-buffer": "~0.1",
-		"shortid": "~2.2",
-		"uuid": "~3.1"
-	},
-	"devDependencies": {
-		"@types/chai": "~4.1",
-		"@types/chai-spies": "~0.0.1",
-		"@types/flake-idgen": "^0.1.30",
-		"@types/mocha": "~5.2",
-		"@types/node": "~8.10",
-		"@types/shortid": "0.0.29",
-		"@types/uuid": "^3.4.4",
-		"chai": "~4.1",
-		"chai-spies": "git://github.com/chaijs/chai-spies.git#b11aeebf",
-		"mocha": "~5.2",
-		"nyc": "~12.0",
-		"rimraf": "~2.6",
-		"ts-declarator": "~1.0",
-		"tslint": "~5.10",
-		"typescript": "~2.9"
-	},
-	"peerDependencies": {
-		"@micro-fleet/common": "~1.2"
-	},
-	"publishConfig": {
-		"access": "public"
-	},
-	"author": {
-		"name": "gennovative",
-		"url": "https://github.com/gennovative"
-	},
-	"engines": {
-		"node": ">=8.0.0"
-	},
-	"license": "MIT",
-	"private": false,
-	"main": "./dist/app/index.js",
-	"types": "./typings/app.d.ts",
-	"repository": "git@github.com:gennovative/micro-fleet-id-generator.git"
+    "name": "@micro-fleet/id-generator",
+    "description": "Provides methods to generate database primary key ID.",
+    "version": "1.0.2",
+    "scripts": {
+        "build": "npm run compile && npm run test",
+        "clean": "rimraf dist coverage .nyc_output",
+        "definition": "node ../../../scripts/definition.js",
+        "dedup": "rimraf node_modules/**/node_modules/**/typings",
+        "compile": "npm run clean && npm run lint && tsc --pretty",
+        "release": "npm run build && npm run definition",
+        "test": "nyc --reporter=json --reporter=html --include dist/app/**/*.js mocha --recursive dist/test",
+        "lint": "tslint -c tslint.json 'src/**/*.ts'",
+        "watch": "npm run clean && npm run lint && tsc --watch --pretty",
+        "linkPackages": "node ../../../scripts/linkPackages.js",
+        "postinstall": "npm run linkPackages",
+        "prepublishOnly": "npm run release"
+    },
+    "dependencies": {
+        "flake-idgen": "~1.1",
+        "shortid": "~2.2",
+        "uuid": "~3.1"
+    },
+    "devDependencies": {
+        "@types/chai": "~4.1",
+        "@types/chai-spies": "~0.0.1",
+        "@types/flake-idgen": "~0.1",
+        "@types/mocha": "~5.2",
+        "@types/node": "~10.14",
+        "@types/shortid": "0.0.29",
+        "@types/uuid": "~3.4",
+        "chai": "~4.1",
+        "chai-spies": "git://github.com/chaijs/chai-spies.git#b11aeebf",
+        "mocha": "~5.2",
+        "nyc": "~13.3",
+        "rimraf": "~2.6",
+        "ts-declarator": "~1.0",
+        "tslint": "~5.11",
+        "typescript": "~3.3"
+    },
+    "peerDependencies": {
+        "@micro-fleet/common": "~1.3"
+    },
+    "publishConfig": {
+        "access": "public"
+    },
+    "author": {
+        "name": "gennovative",
+        "url": "https://github.com/gennovative"
+    },
+    "engines": {
+        "node": ">=10.0.0"
+    },
+    "license": "MIT",
+    "private": false,
+    "main": "./dist/app/index.js",
+    "types": "./typings/app.d.ts",
+    "repository": "git@github.com:gennovative/micro-fleet-id-generator.git"
 }
