@@ -1,3 +1,3 @@
 export class Types {
-	public static readonly ID_PROVIDER = 'id-generator.IdProvider';
+    public static readonly ID_PROVIDER = 'id-generator.IdProvider'
 }
