@@ -1,4 +1,4 @@
-export * from './IdGenerator';
-export * from './IdProviderAddOn';
-export * from './register-addon';
-export * from './Types';
+export * from './IdGenerator'
+export * from './IdProviderAddOn'
+export * from './register-addon'
+export * from './Types'
